+++ lib/src/model/scene_value_validation_node_rect.dart
@@ -6,6 +6,7 @@ import '../core/scene_limits.dart';
 import '../core/nodes.dart';
 import 'scene_value_validation_primitives.dart';
 import 'scene_value_validation_support.dart';
+import 'scene_value_validation_vector_width.dart';
 
 void sceneValidateRectNodeSnapshot(
   RectNodeSnapshot rect, {
@@ -67,16 +68,9 @@ void _sceneValidateRectNodeFields({
     max: sceneSizeMax,
     onError: onError,
   );
-  sceneValidateNonNegativeDouble(
+  sceneValidateNonNegativeVectorWidth(
     strokeWidth,
     field: '$field.strokeWidth',
-    onError: onError,
-  );
-  sceneValidateDoubleInRange(
-    strokeWidth,
-    field: '$field.strokeWidth',
-    min: 0,
-    max: sceneThicknessMax,
     onError: onError,
   );
 }